Dont break when resize_to is not used in the project yet

+++ lib/middleman-simple-thumbnailer/image_store.rb
@@ -24,17 +24,21 @@ module MiddlemanSimpleThumbnailer
     end
 
     def each
-      File.open(@tmp_path, "r") do |f|
-        f.flock(File::LOCK_SH)
-        resized_images = f.size > 0 ? Marshal.load(f) : {}
-        resized_images.values.each do |store_entry|
-          yield *store_entry
+      if File.exist?(@tmp_path)
+        File.open(@tmp_path, "r") do |f|
+          f.flock(File::LOCK_SH)
+          resized_images = f.size > 0 ? Marshal.load(f) : {}
+          resized_images.values.each do |store_entry|
+            yield *store_entry
+          end
         end
       end
     end
 
     def delete
-      File.delete(@tmp_path)
+      if File.exist?(@tmp_path)
+        File.delete(@tmp_path)
+      end
     end
 
   end
